feat(#21): move verifyJWT to it's own module
