+++ src/auth.js
@@ -1,0 +1,14 @@
+const jwt = require('jsonwebtoken')
+
+function verifyJWT(req, res, next){
+  //verify if has token on header
+  var token = req.headers['x-access-token'];
+  if (!token) return res.status(401).send({ auth: false, message: 'No token provided.' });
+
+  jwt.verify(token, 'secretarg', function(err, decoded) {
+    if (err) return res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
+    next();
+  });
+}
+
+module.exports = verifyJWT
