+++ src/server.js
@@ -3,7 +3,6 @@
 const mongoose = require('mongoose')
 const EventEmitter = require('events')
 const express = require('express')
-const jwt = require('jsonwebtoken')
 const bodyParser = require('body-parser')
 
 const WebProxy = require('../lib/webProxy.js')
@@ -11,6 +10,7 @@ const Service = require('../lib/service.js')
 const startMongo = require('../config/mongo.js').startMongo
 const RegisterEmitter = require('../events/registerEmitter.js')
 const LogEmitter = require('../events/logEmitter.js')
+const verifyJWT = require('./auth.js')
 
 const { PORT } = process.env
 const { KEY } = process.env
@@ -115,17 +115,6 @@ app.post('/register', (req, res, next) =>   {
 
   res.status(200).send({'message': 'Event received'})
 })
-
-function verifyJWT(req, res, next){
-  //verify if has token on header
-  var token = req.headers['x-access-token'];
-  if (!token) return res.status(401).send({ auth: false, message: 'No token provided.' });
-
-  jwt.verify(token, 'secretarg', function(err, decoded) {
-    if (err) return res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
-    next();
-  });
-}
 
 app.listen(PORT, async () => {
     try {
